+++ packages/web/src/javascripts/Components/Preferences/PreferencesMenu.ts
@@ -69,7 +69,7 @@ export class PreferencesMenu {
   }
 
   private loadLatestVersions(): void {
-    PackageProvider.load(this.application)
+    PackageProvider.load()
       .then((versions) => {
         if (versions) {
           this._extensionLatestVersions = versions