+++ packages/api/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.14](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.13...@standardnotes/api@1.26.14) (2023-06-25)
+
+### Bug Fixes
+
+* fixes issue where a renewed session would in some cases would not save to disk, causing a reauthentication prompt to occur ([bb804ff](https://github.com/standardnotes/app/commit/bb804fff45328eef3c66ec3a48dc000c21e37020))
+
+## [1.26.13](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.12...@standardnotes/api@1.26.13) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/api
+
+## [1.26.12](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.11...@standardnotes/api@1.26.12) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/api
+
+## [1.26.11](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.10...@standardnotes/api@1.26.11) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/api
+
 ## [1.26.10](https://github.com/standardnotes/app/compare/@standardnotes/api@1.26.9...@standardnotes/api@1.26.10) (2023-05-22)
 
 **Note:** Version bump only for package @standardnotes/api