+++ packages/clipper/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.51](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.50...@standardnotes/clipper@1.1.51) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.50](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.49...@standardnotes/clipper@1.1.50) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.49](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.48...@standardnotes/clipper@1.1.49) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.48](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.47...@standardnotes/clipper@1.1.48) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.47](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.46...@standardnotes/clipper@1.1.47) (2023-06-23)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
 ## [1.1.46](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.45...@standardnotes/clipper@1.1.46) (2023-06-16)
 
 **Note:** Version bump only for package @standardnotes/clipper