+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.53.60](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.53.59...@standardnotes/mobile@3.53.60) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.53.59](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.53.58...@standardnotes/mobile@3.53.59) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.53.58](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.53.57...@standardnotes/mobile@3.53.58) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.53.57](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.53.56...@standardnotes/mobile@3.53.57) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.53.56](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.53.55...@standardnotes/mobile@3.53.56) (2023-06-23)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.53.55](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.53.54...@standardnotes/mobile@3.53.55) (2023-06-16)
 
 **Note:** Version bump only for package @standardnotes/mobile