+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.328](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.327...@standardnotes/releases@1.4.328) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.327](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.326...@standardnotes/releases@1.4.327) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.326](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.325...@standardnotes/releases@1.4.326) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.325](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.324...@standardnotes/releases@1.4.325) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.324](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.323...@standardnotes/releases@1.4.324) (2023-06-23)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.4.323](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.322...@standardnotes/releases@1.4.323) (2023-06-16)
 
 **Note:** Version bump only for package @standardnotes/releases