+++ packages/services/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.60.3](https://github.com/standardnotes/app/compare/@standardnotes/services@1.60.2...@standardnotes/services@1.60.3) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/services
+
+## [1.60.2](https://github.com/standardnotes/app/compare/@standardnotes/services@1.60.1...@standardnotes/services@1.60.2) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/services
+
+## [1.60.1](https://github.com/standardnotes/app/compare/@standardnotes/services@1.60.0...@standardnotes/services@1.60.1) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/services
+
+# [1.60.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.59.12...@standardnotes/services@1.60.0) (2023-06-25)
+
+### Features
+
+* Added a conflict resolution dialog and a Conflicts view for easier management of conflicts ([#2337](https://github.com/standardnotes/app/issues/2337)) ([e0e9249](https://github.com/standardnotes/app/commit/e0e92493342b18560e7e55ece22518e738824df0))
+
 ## [1.59.12](https://github.com/standardnotes/app/compare/@standardnotes/services@1.59.11...@standardnotes/services@1.59.12) (2023-06-06)
 
 **Note:** Version bump only for package @standardnotes/services