+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.17.2](https://github.com/standardnotes/app/compare/@standardnotes/utils@1.17.1...@standardnotes/utils@1.17.2) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/utils
+
+## [1.17.1](https://github.com/standardnotes/app/compare/@standardnotes/utils@1.17.0...@standardnotes/utils@1.17.1) (2023-06-25)
+
+**Note:** Version bump only for package @standardnotes/utils
+
+# [1.17.0](https://github.com/standardnotes/app/compare/@standardnotes/utils@1.16.5...@standardnotes/utils@1.17.0) (2023-06-25)
+
+### Features
+
+* Added a conflict resolution dialog and a Conflicts view for easier management of conflicts ([#2337](https://github.com/standardnotes/app/issues/2337)) ([e0e9249](https://github.com/standardnotes/app/commit/e0e92493342b18560e7e55ece22518e738824df0))
+
 ## [1.16.5](https://github.com/standardnotes/app/compare/@standardnotes/utils@1.16.4...@standardnotes/utils@1.16.5) (2023-03-01)
 
 **Note:** Version bump only for package @standardnotes/utils